Add ability to track over thread ID's when gdb is used with rr: https://rr-project.org/
This commit enables the ability to use gdb with rr and rr in eclipse.
Fixes Bug 23454 on Bugzilla
When replaying a thread creation gdb increments the thread ID number
shown with 'info threads'. This caused different thread ID's,
while replaying the same thread exectuion.
In thread.c, we made a change that looks, if this thread already
got a number and if yes, the thread gets the old number assigned.
When using in eclipse, we have to manually increment the
'highest_thread_num' once. Therefore the command:
'rr_running_under_eclipse' was added, which can be used 'once'
to set the correct highest_thread_num.

gdb/ChangeLog:

        * thread.c: Add command